added isWebpSupported(), bindings to JS and java, test, and usage in examples.

+++ docs_src/src_raw/viewer/filament-viewer.js
@@ -318,7 +318,9 @@ class FilamentViewer extends LitElement {
                         resourceLoader.delete();
                         stbProvider.delete();
                         ktx2Provider.delete();
-                        webpProvider.delete();
+                        if (webpProvider) {
+                            webpProvider.delete();
+                        }
                         this.animator = this.asset.getInstance().getAnimator();
                         this.animationStartTime = Date.now();
                     }
@@ -336,12 +338,17 @@ class FilamentViewer extends LitElement {
 
                 const stbProvider = new Filament.gltfio$StbProvider(this.engine);
                 const ktx2Provider = new Filament.gltfio$Ktx2Provider(this.engine);
-                const webpProvider = new Filament.gltfio$WebpProvider(this.engine);
+                let webpProvider = null;
 
                 resourceLoader.addStbProvider("image/jpeg", stbProvider);
                 resourceLoader.addStbProvider("image/png", stbProvider);
                 resourceLoader.addKtx2Provider("image/ktx2", ktx2Provider);
-                resourceLoader.addWebpProvider("image/webp", webpProvider);
+
+
+                if (Filament.gltfio$WebpProvider.isWebpSupported()) {
+                    webpProvider = new Filament.gltfio$WebpProvider(this.engine);
+                    resourceLoader.addWebpProvider("image/webp", webpProvider);
+                }
 
                 let remaining = Object.keys(this.srcBlobResources).length;
                 for (const name in this.srcBlobResources) {
